Inline Help: move video dialog to separate async-loaded component
