+++ client/blocks/inline-help/dialog.scss
@@ -1,0 +1,22 @@
+.inline-help__richresult__dialog.dialog.card {
+	width: 80vw;
+	max-width: 850px;
+}
+
+.inline-help__richresult__dialog__video {
+	position: relative;
+	padding-bottom: 56.25%;
+	padding-top: 30px;
+	height: 0;
+	overflow: hidden;
+
+	iframe,
+	object,
+	embed {
+		position: absolute;
+		top: 0;
+		left: 0;
+		width: 100%;
+		height: 100%;
+	}
+}
