+++ client/blocks/inline-help/index.jsx
@@ -19,8 +19,6 @@ import { recordTracksEvent } from 'state/analytics/actions';
 import getGlobalKeyboardShortcuts from 'lib/keyboard-shortcuts/global';
 import Button from 'components/button';
 import HappychatButton from 'components/happychat/button';
-import Dialog from 'components/dialog';
-import ResizableIframe from 'components/resizable-iframe';
 import isHappychatOpen from 'state/happychat/selectors/is-happychat-open';
 import hasActiveHappychatSession from 'state/happychat/selectors/has-active-happychat-session';
 import AsyncLoad from 'components/async-load';
@@ -39,6 +37,10 @@ const debug = debugFactory( 'calypso:inline-help' );
 
 const InlineHelpPopover = props => (
 	<AsyncLoad { ...props } require="blocks/inline-help/popover" placeholder={ null } />
+);
+
+const InlineHelpDialog = props => (
+	<AsyncLoad { ...props } require="blocks/inline-help/dialog" placeholder={ null } />
 );
 
 class InlineHelp extends Component {
@@ -123,19 +125,6 @@ class InlineHelp extends Component {
 
 	closeDialog = () => this.setState( { showDialog: false } );
 
-	getDialogButtons() {
-		const { translate } = this.props;
-		const { dialogType } = this.state;
-
-		if ( dialogType === 'video' ) {
-			return [
-				<Button onClick={ this.closeDialog }>{ translate( 'Close', { textOnly: true } ) }</Button>,
-			];
-		}
-
-		return [];
-	}
-
 	setNotification = status => {
 		this.setState( { showChecklistNotification: status } );
 	};
@@ -152,13 +141,6 @@ class InlineHelp extends Component {
 			'is-active': isPopoverVisible,
 			'has-notification': showChecklistNotification,
 		};
-
-		/* @TODO: This class is not valid and this tricks the linter
-		 		  fix this class and fix the linter to catch similar instances.
-		 */
-		const iframeClasses = classNames( 'inline-help__richresult__dialog__video' );
-		const dialogClasses = classNames( 'inline-help__richresult__dialog', dialogType );
-		const dialogButtons = this.getDialogButtons();
 
 		return (
 			<div className="inline-help">
@@ -189,27 +171,11 @@ class InlineHelp extends Component {
 					/>
 				) }
 				{ showDialog && (
-					<Dialog
-						additionalClassNames={ dialogClasses }
-						isVisible
-						buttons={ dialogButtons }
-						onCancel={ this.closeDialog }
+					<InlineHelpDialog
+						dialogType={ dialogType }
+						videoLink={ videoLink }
 						onClose={ this.closeDialog }
-					>
-						{ dialogType === 'video' && (
-							<div className={ iframeClasses }>
-								<ResizableIframe
-									src={ videoLink + '?rel=0&amp;showinfo=0&amp;autoplay=1' }
-									frameBorder="0"
-									seamless
-									allowFullScreen
-									autoPlay
-									width="640"
-									height="360"
-								/>
-							</div>
-						) }
-					</Dialog>
+					/>
 				) }
 				{ this.props.isHappychatButtonVisible && config.isEnabled( 'happychat' ) && (
 					<HappychatButton className="inline-help__happychat-button" allowMobileRedirect />
